fixed errors

+++ ftpservice.h
@@ -59,7 +59,7 @@ typedef struct cmd_map_s {
 } cmd_map_t;
 
 extern char root_directory[PATH_LEN];
-extern uint8_t hostip_octets[4];
+extern int hostip_octets[4];
 extern cmd_map_t cmd_map[NUM_CMDS];
 
 void *handle_session(void *clientfd);
